Extract find_index module from accessor_general
Here I just move the code without changes.
